requirements.txt
python-dotenv~=1.0.0
diffusers~=0.15.1
Pillow~=9.5.0
numpy~=1.24.2
lit~=16.0.1
pip~=21.3.1
wheel~=0.37.1
requests~=2.28.2
PyYAML~=6.0
setuptools~=60.2.0
regex~=2023.3.23
sympy~=1.11.1
mpmath~=1.3.0
tqdm~=4.65.0
urllib3~=1.26.15
triton~=2.0.0
filelock~=3.12.0
certifi~=2022.12.7
idna~=3.4
packaging~=23.1
zipp~=3.15.0
torch~=2.0.0
Jinja2~=3.1.2
networkx~=3.1
transformers~=4.28.1
tokenizers~=0.13.3
psutil~=5.9.5
accelerate~=0.18.0
pika~=1.3.1